// src/app/interfaces/configuration-response.interface.ts
// Generated by https://quicktype.io

export interface ConfigurationResponse {
  projects:     DetailButton[];
  builds:       DetailButton[];
  packaging:    DetailButton[];
  domains:      DetailButton[];
  bootVersions: DetailButton[];
  languages:    DetailButton[];
  javaVersions: DetailButton[];
  groupId:      string;
  artifactId:   string;
  version:      string;
  name:         string;
  description:  string;
  packageName:  string;
}

export interface DetailButton {
  id:          string;
  name:        string;
  checked:     boolean;
  def:         boolean;
  description?: string;
  action?:      string;
  esGinni:     boolean;
  disabled?:   boolean;
}

export interface ConfigurationResponseV2 {
  configurations: ConfigurationResponse;
  dependencies: any;
}